Introduce file header checking script
This change introduces a script for checking whether files passed to it
contain the appropriate header. A valid header is either an (for the
moment) arbitrary shebang or a comment containing the file's file name.
If no such header is found a message will be printed and an error code
returned.